Allow tenant networks to be shared with domain 0.
If domain 0 has an IP address on a given bridge, then add a flow rule to
allow traffic to leave that port.  This allows for the case where domain 0
and the tenant network are sharing a bridge, which is useful in non-production
environments.

At the same time, simplify the interface to ovs_configure_base_rules.py.
There is no need to pass the bridge in, because it's implicit in the selection
of the physical interface, and can be obtained using ovs-vsctl iface-to-br.

Having made these two changes, it's now reasonable to apply flow rules to all
interfaces as a default (if not overridden in the sysconfig file).

Change-Id: I2a33ed55246d49b0e4f57db909e1b40351d27602