Bug #886353: Faults raised by OpenStack API Resource handlers fail to be reported properly
In Resource.__call__, catch any faults.Fault that are thrown by the request
handler, and make those the action_result.  This means that the Fault doesn't
fall into the exception handler below (faults.Fault is a subclass of
webob.exc.HTTPException) and so we don't attempt to wrap a Fault in a Fault.

Change-Id: Ica31723ddd621332ec4be75db0d675bc98905e43