Cherry-pick the below.  In dasara, we don't have the _set_vcpu_weight code, so there's nothing to fix there, and the spawn signature is different so that needed fixing up.
XenAPI: destroy VM if the post-boot operations fail.

If the post-boot operations fail (i.e. file injection, admin password
setting, network setup, vCPU weight setting) then destroy the VM.

This may seem a bit aggressive, but the VM is likely to be unreachable
and/or unusable by the tenant, so it seems like the best thing to do.  The
alternative would be to show this failure as a task_state, but then you're
going to have to set appropriate billing policy if you don't want the
customer to be charged in this scenario.  I'm open to other suggestions.

Doing this has exposed the fact that our unit tests don't cover any of
this stage.  We were failing in various ways during unit testing, but none
of the tests noticed because there was no external indication of the failure.
Cleaning up the VM meant that I had to fix _test_spawn and _create_instance
to set injected_files and admin_pass before calling xenapi_conn.spawn.  These
two fields are set by compute.manager._run_instance before passing the
spawn request to the virt layer.

I also had to fix _set_vcpu_weight to cope with the absence of a
vcpu_weight value on the instance_type DB object.

Change-Id: I30cfc23c363939107fdce980a59a25395d858e10

Conflicts:

	nova/tests/test_xenapi.py
	nova/virt/xenapi/vmops.py