changeset:   319:a2668e8c503d user:        Ewan Mellor <ewan.mellor@citrix.com> date:        Tue Nov 29 18:36:33 2011 -0800 files:       upstream/nova/utils.py upstream/nova/volume/driver.py upstream/tool s/pip-requires description: OS-737: Move all OpenStack service to use a Host GUID OS-824: nova-volume is trying to send iscsiadm to a MAC address
Change ISCSIDriver to set volume.provider_location during create_export.
This records the location of the LUN, so that nova-compute does not need
to run the iSCSI discovery code itself.

As part of this, include the IP address of the target in the provider_location.
This IP address is carefully selected to be the one on the correct network,
i.e. matching --iscsi_ip_prefix.  This means that we don't use volume's
host identifier (which could be an opaque ID) when trying to connect to the
iSCSI target -- we use the correct IP address instead.

The string-join to set provider_location is shared with ZadaraBEDriver,
which was doing something similar.  I've brought that into a helper function.

This change includes whataremyips(), taken from swift.common.utils.

This adds a dependency on netifaces.