bug 921087: i18n-key and local-storage hard-coded in xenapi
This fix introduces a new flag 'sr_matching_filter', whose default value is
'other-config:i18n-key=local-storage'. This filter is used for finding the
SR on which to install guest instances. The default value is the Local
Storage in default XenServer/XCP installations, and it is what was hard-coded
so far.

To select an SR with a different matching criteria, this flag can be set
to 'other-config:my_favorite_sr=true'. On the other hand, to fall back on
the Default SR, as displayed by XenCenter and as returned by
xenapi.pool.get_default_SR, this flag can be set to to 'default-sr:true'.

This changeset also makes a small code simplification along the way.

Change-Id: Ia5ee438389c59a5ef0b858e8548643d57ef16c77