xapi: fix create hypervisor pool
Fixes bug 1049099.

Fixing problems with the rpc api when creating hypervisor pools with
xenapi. Rpc calls were not using the compute_rpcapi approach, thus were
not properly versioned. Apart from that as the slave parameters were not
forwarded to the master. Separate RPC calls are introduced for those
execution paths.

Change-Id: Ie44a1c09ef204affc4a657c71557691e83b22c22